Macro optimization for HashMap
`HashMap::entry()` requires a move of the key. When adding a new entry
is rare, cloning key objects for every call of this method may be not a
good option. Even for small String keys (we could expect many symbol
names are not so long) it invokes heap allocation due to the lack of
small string optimization in Rust.

Instead, we can use `HashMap::get()` or `HashMap::get_mut()` to access
most likely existing values.

Cons: For a new entry, hash search is performed twice (unless the
compiler is super smart).