Made dollar variables local by default.
During execution, when a statement creates multiple terms,
each term will now have their own independent dollar variable.
This means that changes in the execution of one branch do not
affect the other branch.

To enable the old non-local behaviour, use the NonLocal
attribute. See examples/nonlocal.frm for an example.

This does not affect the behaviour that for each term in the input
the dollar variable keeps its previous value. To make sure each
input term starts with the same value, place an assignment at the
start of the module instead of in the global scope.